+++ src/constants/index.ts
@@ -3,3 +3,4 @@ export * from './sheets-name/sheets-name';
 export * from './sheets-name/sheets-name.enum';
 export * from './dates-info/dates-info';
 export * from './restaurant-data/restaurant-data';
+export * from './responses';